+++ CHANGELOG.md
@@ -19,7 +19,9 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - Added `cu::DeviceMemory::memset2D()` and `cu::Stream::memset2DAsync()`
 - Added `cufft::FFT1DR2C` and `cufft::FFT1DC2R`
 - Added `cu::Device::getOrdinal()`
-- Improved HIP documentation and building instructions in `README.md`
+- Added deprecated warning to `cu::Context` constructor
+- Improved HIP documentation and building instructions
+
 
 ### Changed