+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Changed
+
+- Bugfix `cu::DeviceMemory` constructors with `CUdeviceptr` argument.
+
 ## [0.9.0] - 2025-03-18
 
 ### Added