+++ CHANGELOG.md
@@ -11,6 +11,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- Reimplementation of `inline_local_includes` with one CMake file
 - Changed `cu::Device` constructor to accept `unsigned int` instead of `int`
 
 ## [0.9.0] - 2025-03-18