+++ CHANGELOG.md
@@ -9,6 +9,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 - Added `nvrtc::addNameExpression` and `nvrtc::getLoweredName`
 - Added header file to ease the use of embedded kernel sources
+- Add `CUDAWRAPPERS_COMPONENTS` CMake option
 
 ### Changed