+++ CHANGELOG.md
@@ -9,7 +9,8 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 - Added `nvrtc::addNameExpression` and `nvrtc::getLoweredName`
 - Added header file to ease the use of embedded kernel sources
-- Add `CUDAWRAPPERS_COMPONENTS` CMake option
+- Added `CUDAWRAPPERS_COMPONENTS` CMake option
+- Added `cu::Stream::launchHostFunc`
 
 ### Changed